# arch/api.yaml
openapi: 3.0.3
...
info:
  title: Moelasware
  description: ''
  version: 1.0.11
servers:
  - url: https://moelasware.io
paths:
...
  /quizzes:
  # TODO: ADD query parameters to filter out quizzes
  # by user, for example.
    description: Questions with answers
    post:
      description: Post a quiz to be reviewed
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/quiz"
            examples:
              quiz_example:
                $ref: "#/components/examples/quiz_example"
              quiz_example_2:
                $ref: "#/components/examples/quiz_example_2"
            
      responses:
        '201':
            description: "Created a new quiz successfully"
            content:
              application/json:
                schema:
                  type: integer
                
                examples:
                  return_id:
                    value:
                      id: 3
    get:
      description: "Get all quizzes"
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: "Retrieved quizzes successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/quiz"
...
components:
  schemas:
    id:
      type: integer
      minimum: 1
...
    quiz:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
        question: 
          type: string
          example: "What color is an orange?"
        
        author: 
          $ref: '#/components/schemas/id'

        reviews:
          type: array
          items:
              $ref: '#/components/schemas/review'

        description:
          type: string
          example: "This question is about the fruit, orange, and it's color"
          
        tags:
          type: array
          items: 
            type: string
          example:
            ["Animals", "Names"]
          
        accepted:
          type: boolean

        answers:
          type: array
          items:
            $ref: '#/components/schemas/quiz_answer'

      required:
        - question
        - answers
        - accepted
    quiz_answer:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
        text:
          type: string
        correct:
          type: boolean
        justification:
          type: string
          
      example:
        text: "An orange is red"
        correct: false
        justification: "Oranges are orange, not red"
    review:
      properties:
        id:
          $ref: '#/components/schemas/id'
        reviewer: 
          $ref: "#/components/schemas/id"
        accepted:
          type: boolean
          example: false
        comment:
          type: string
          example: "There is no correct answer"
  parameters:
    offsetParam:  
      name: offset
      in: query
      description: Number of items to skip before returning the results.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    limitParam:       
      name: limit
      in: query
      description: Maximum number of items to return.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
  examples:
    quiz_example:
      value:
        question: "What color is an orange?"
        accepted: true
        answers:  
          - text: "Red"
            correct: false
            justification: "Oranges are orange, not red!"
          - text: "Orange"
            correct: true
            justification: "Oranges are orange!"
          - text: "Blue"
            correct: false
            justification: "Oranges are orange, not blue!"
          - text: "Green"
            correct: false
            justification: "Oranges are orange, not green!"
      summary: "Orange"
    quiz_example_2:
      value:
        question: "Which of the following startup names is objectively superior?"
        accepted: false
        reviews: "TODO: ADD REVIEWS"
        answers:  
          - text: "Moelasware"
            correct: false
            justification: "Doesn't even have quiz in the name"
          - text: "Quizzly"
            correct: false
            justification: ""
          - text: "Quizzhut"
            correct: true
      summary: "Quizzhut"
...